:sparkles: Use Flux to add authentication
With a Flux a stream of requests can be decorated, allowing this
transformation in a wider range of use-cases.

Note that the combination of flatMap and map achieves a similar result
as zipWith, but blocks the completed-signal from the Mono in
validUserToken and repeats these calls instead.

Future implementations may switch to a Flux for the UserToken which
always returns a valid token on request and renews the tokens when
necessary.