:hammer: Condense discovery into one class and move to main package
The discovery result can be a record and is only ever used in the discovery
method, so it can be a private member of the discovery class.

Since there is only one class left for discovery, it does not need its own
package anymore.

Get the HttpClient from injection instead of constructor.

Make the discovery Bean a Singleton.